refactor: add remaining indexer tests
