+++ crates/iota-json-rpc-tests/tests/indexer_api.rs
@@ -3,23 +3,36 @@
 
 #[cfg(not(msim))]
 use std::str::FromStr;
+use std::time::SystemTime;
 
+use futures::stream::StreamExt;
+use iota_json::{call_args, type_args};
 use iota_json_rpc_api::IndexerApiClient;
 use iota_json_rpc_types::{
-    IotaObjectDataFilter, IotaObjectDataOptions, IotaObjectResponseQuery, ObjectsPage,
+    EventFilter, EventPage, IotaObjectDataFilter, IotaObjectDataOptions, IotaObjectResponseQuery,
+    IotaTransactionBlockResponse, IotaTransactionBlockResponseOptions,
+    IotaTransactionBlockResponseQuery, ObjectsPage, TransactionFilter,
 };
 use iota_macros::sim_test;
 use iota_protocol_config::ProtocolConfig;
+use iota_sdk::IotaClientBuilder;
 use iota_swarm_config::genesis_config::AccountConfig;
+use iota_test_transaction_builder::TestTransactionBuilder;
 use iota_types::{
+    IOTA_FRAMEWORK_ADDRESS,
     base_types::{IotaAddress, MoveObjectType, ObjectID},
     collection_types::VecMap,
     crypto::deterministic_random_account_key,
     digests::TransactionDigest,
+    gas_coin::GAS,
     id::UID,
     object::{Data, MoveObject, OBJECT_START_VERSION, ObjectInner, Owner},
+    programmable_transaction_builder::ProgrammableTransactionBuilder,
+    quorum_driver_types::ExecuteTransactionRequestType,
     stardust::output::{Irc27Metadata, Nft},
+    transaction::{Command, ObjectArg, SenderSignedData, TransactionData},
 };
+use move_core_types::{identifier::Identifier, language_storage::TypeTag};
 use test_cluster::TestClusterBuilder;
 
 #[sim_test]
@@ -128,6 +141,510 @@ async fn test_nft_display_object() -> Result<(), anyhow::Error> {
     assert_eq!(nft_display["media_type"], "media type value");
     assert_eq!(nft_display["name"], "name value");
     assert_eq!(nft_display["version"], "version value");
+
+    Ok(())
+}
+
+#[sim_test]
+async fn query_events_no_events_descending() {
+    let cluster = TestClusterBuilder::new().build().await;
+
+    let client = cluster.rpc_client();
+
+    let indexer_events = client
+        .query_events(
+            EventFilter::Sender(IotaAddress::random_for_testing_only()),
+            None,
+            None,
+            Some(true),
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(indexer_events, EventPage::empty())
+}
+
+#[sim_test]
+async fn query_events_no_events_ascending() {
+    let cluster = TestClusterBuilder::new().build().await;
+
+    let client = cluster.rpc_client();
+
+    let indexer_events = client
+        .query_events(
+            EventFilter::Sender(IotaAddress::random_for_testing_only()),
+            None,
+            None,
+            None,
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(indexer_events, EventPage::empty())
+}
+
+#[ignore]
+#[sim_test]
+async fn query_events_unsupported_events() {
+    let cluster = TestClusterBuilder::new()
+        .with_epoch_duration_ms(1000)
+        .build()
+        .await;
+
+    tokio::time::sleep(tokio::time::Duration::from_millis(10000)).await;
+
+    let client = cluster.rpc_client();
+
+    // Get the current time in milliseconds since the UNIX epoch
+    let now_millis = SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .unwrap()
+        .as_millis();
+
+    // Subtract 10 minutes from the current time
+    let ten_minutes_ago = now_millis - (10 * 60 * 1000); // 600 seconds = 10 minutes
+
+    let unsupported_filters = vec![
+        EventFilter::All(vec![]),
+        EventFilter::Any(vec![]),
+        EventFilter::And(
+            Box::new(EventFilter::Any(vec![])),
+            Box::new(EventFilter::Any(vec![])),
+        ),
+        EventFilter::Or(
+            Box::new(EventFilter::Any(vec![])),
+            Box::new(EventFilter::Any(vec![])),
+        ),
+        EventFilter::TimeRange {
+            start_time: ten_minutes_ago as u64,
+            end_time: now_millis as u64,
+        },
+        EventFilter::MoveEventField {
+            path: String::default(),
+            value: serde_json::Value::Bool(true),
+        },
+    ];
+
+    for event_filter in unsupported_filters {
+        let result = client.query_events(event_filter, None, None, None).await;
+
+        let err = result.unwrap_err();
+        let msg = err.to_string();
+
+        assert_eq!(
+            msg,
+            "Indexer does not support the feature with error: `This type of EventFilter is not supported.`"
+        );
+    }
+}
+
+#[ignore]
+#[sim_test]
+async fn query_events_supported_events() {
+    let cluster = TestClusterBuilder::new().build().await;
+
+    tokio::time::sleep(tokio::time::Duration::from_millis(10000)).await;
+
+    let client = cluster.rpc_client();
+
+    let supported_filters = vec![
+        EventFilter::Sender(IotaAddress::ZERO),
+        EventFilter::Transaction(TransactionDigest::ZERO),
+        EventFilter::Package(ObjectID::ZERO),
+        EventFilter::MoveEventModule {
+            package: ObjectID::ZERO,
+            module: "x".parse().unwrap(),
+        },
+        EventFilter::MoveEventType("0xabcd::MyModule::Foo".parse().unwrap()),
+        EventFilter::MoveModule {
+            package: ObjectID::ZERO,
+            module: "x".parse().unwrap(),
+        },
+    ];
+
+    for event_filter in supported_filters {
+        let result = client.query_events(event_filter, None, None, None).await;
+        assert!(result.is_ok());
+    }
+}
+
+#[sim_test]
+async fn test_get_owned_objects() -> Result<(), anyhow::Error> {
+    let cluster = TestClusterBuilder::new().build().await;
+
+    let http_client = cluster.rpc_client();
+    let address = cluster.get_address_0();
+
+    let objects = http_client
+        .get_owned_objects(
+            address,
+            Some(IotaObjectResponseQuery::new_with_options(
+                IotaObjectDataOptions::new(),
+            )),
+            None,
+            None,
+        )
+        .await?;
+    assert_eq!(5, objects.data.len());
+
+    Ok(())
+}
+
+#[sim_test]
+async fn test_query_transaction_blocks_pagination() -> Result<(), anyhow::Error> {
+    let cluster = TestClusterBuilder::new().build().await;
+
+    let context = &cluster.wallet;
+
+    let mut tx_responses: Vec<IotaTransactionBlockResponse> = Vec::new();
+
+    let client = context.get_client().await.unwrap();
+
+    for address in cluster.get_addresses() {
+        let objects = client
+            .read_api()
+            .get_owned_objects(
+                address,
+                Some(IotaObjectResponseQuery::new_with_options(
+                    IotaObjectDataOptions::new()
+                        .with_type()
+                        .with_owner()
+                        .with_previous_transaction(),
+                )),
+                None,
+                None,
+            )
+            .await?
+            .data;
+        let gas_id = objects.last().unwrap().object().unwrap().object_id;
+
+        assert_eq!(5, objects.len());
+
+        // Make some transactions
+        for obj in &objects[..objects.len() - 1] {
+            let oref = obj.object().unwrap();
+            let data = client
+                .transaction_builder()
+                .transfer_object(address, oref.object_id, Some(gas_id), 1_000_000, address)
+                .await?;
+            let tx = cluster.wallet.sign_transaction(&data);
+
+            let response = client
+                .quorum_driver_api()
+                .execute_transaction_block(
+                    tx,
+                    IotaTransactionBlockResponseOptions::new(),
+                    Some(ExecuteTransactionRequestType::WaitForLocalExecution),
+                )
+                .await
+                .unwrap();
+
+            tx_responses.push(response);
+        }
+    }
+
+    // test get_recent_transactions with smaller range with address filter
+    let query = IotaTransactionBlockResponseQuery {
+        options: Some(IotaTransactionBlockResponseOptions {
+            show_input: true,
+            show_effects: true,
+            show_events: true,
+            ..Default::default()
+        }),
+        filter: Some(TransactionFilter::FromAddress(cluster.get_address_0())),
+    };
+
+    let tx = client
+        .read_api()
+        .query_transaction_blocks(query.clone(), None, Some(3), true)
+        .await
+        .unwrap();
+    assert_eq!(3, tx.data.len());
+    assert!(tx.data[0].transaction.is_some());
+    assert!(tx.data[0].effects.is_some());
+    assert!(tx.data[0].events.is_some());
+    assert!(tx.has_next_page);
+
+    // Read the next page for the last transaction
+    let next_page = client
+        .read_api()
+        .query_transaction_blocks(query, tx.next_cursor, None, true)
+        .await
+        .unwrap();
+
+    assert_eq!(1, next_page.data.len());
+
+    // test get all transactions paged without address filter
+    let first_page = client
+        .read_api()
+        .query_transaction_blocks(
+            IotaTransactionBlockResponseQuery::default(),
+            None,
+            Some(5),
+            false,
+        )
+        .await
+        .unwrap();
+    assert_eq!(5, first_page.data.len());
+    assert!(first_page.has_next_page);
+
+    let second_page = client
+        .read_api()
+        .query_transaction_blocks(
+            IotaTransactionBlockResponseQuery::default(),
+            first_page.next_cursor,
+            None,
+            false,
+        )
+        .await
+        .unwrap();
+    assert!(second_page.data.len() > 5);
+
+    let mut all_txs = first_page.data.clone();
+    all_txs.extend(second_page.data);
+
+    // test get 10 transactions paged
+    let latest = client
+        .read_api()
+        .query_transaction_blocks(
+            IotaTransactionBlockResponseQuery::default(),
+            None,
+            Some(10),
+            false,
+        )
+        .await
+        .unwrap();
+    assert_eq!(10, latest.data.len());
+    assert_eq!(Some(all_txs[9].digest), latest.next_cursor);
+    assert_eq!(all_txs[0..10], latest.data);
+    assert!(latest.has_next_page);
+
+    // test get from address txs in ascending order
+    let address_txs_asc = client
+        .read_api()
+        .query_transaction_blocks(
+            IotaTransactionBlockResponseQuery::new_with_filter(TransactionFilter::FromAddress(
+                cluster.get_address_0(),
+            )),
+            None,
+            None,
+            false,
+        )
+        .await
+        .unwrap();
+    assert_eq!(4, address_txs_asc.data.len());
+
+    // test get from address txs in descending order
+    let address_txs_desc = client
+        .read_api()
+        .query_transaction_blocks(
+            IotaTransactionBlockResponseQuery::new_with_filter(TransactionFilter::FromAddress(
+                cluster.get_address_0(),
+            )),
+            None,
+            None,
+            true,
+        )
+        .await
+        .unwrap();
+    assert_eq!(4, address_txs_desc.data.len());
+
+    // test get from address txs in both ordering are the same.
+    let mut data_asc = address_txs_asc.data;
+    data_asc.reverse();
+    assert_eq!(data_asc, address_txs_desc.data);
+
+    // test get_recent_transactions
+    let tx = client
+        .read_api()
+        .query_transaction_blocks(
+            IotaTransactionBlockResponseQuery::default(),
+            None,
+            Some(20),
+            true,
+        )
+        .await
+        .unwrap();
+    assert_eq!(20, tx.data.len());
+
+    Ok(())
+}
+
+#[sim_test]
+async fn test_query_transaction_blocks() -> Result<(), anyhow::Error> {
+    let mut cluster = TestClusterBuilder::new().build().await;
+    let context = &cluster.wallet;
+    let client = context.get_client().await.unwrap();
+
+    let address = cluster.get_address_0();
+    let objects = client
+        .read_api()
+        .get_owned_objects(
+            address,
+            Some(IotaObjectResponseQuery::new_with_options(
+                IotaObjectDataOptions::new()
+                    .with_type()
+                    .with_owner()
+                    .with_previous_transaction(),
+            )),
+            None,
+            None,
+        )
+        .await?
+        .data;
+
+    // make 2 move calls of same package & module, but different functions
+    let package_id = ObjectID::new(IOTA_FRAMEWORK_ADDRESS.into_bytes());
+    let coin = objects.first().unwrap();
+    let coin_2 = &objects[1];
+    let signer = cluster.wallet.active_address().unwrap();
+
+    let tx_builder = client.transaction_builder().clone();
+    let mut pt_builder = ProgrammableTransactionBuilder::new();
+    let gas = objects.last().unwrap().object().unwrap().object_ref();
+
+    let module = Identifier::from_str("pay")?;
+    let function_1 = Identifier::from_str("split")?;
+    let function_2 = Identifier::from_str("divide_and_keep")?;
+
+    let iota_type_args = type_args![GAS::type_tag()]?;
+    let type_args = iota_type_args
+        .into_iter()
+        .map(|ty| ty.try_into())
+        .collect::<Result<Vec<_>, _>>()?;
+
+    let iota_call_args_1 = call_args!(coin.data.clone().unwrap().object_id, 10)?;
+    let call_args_1 = tx_builder
+        .resolve_and_checks_json_args(
+            &mut pt_builder,
+            package_id,
+            &module,
+            &function_1,
+            &type_args,
+            iota_call_args_1,
+        )
+        .await?;
+    let cmd_1 = Command::move_call(
+        package_id,
+        module.clone(),
+        function_1,
+        type_args.clone(),
+        call_args_1.clone(),
+    );
+
+    let iota_call_args_2 = call_args!(coin_2.data.clone().unwrap().object_id, 10)?;
+    let call_args_2 = tx_builder
+        .resolve_and_checks_json_args(
+            &mut pt_builder,
+            package_id,
+            &module,
+            &function_2,
+            &type_args,
+            iota_call_args_2,
+        )
+        .await?;
+    let cmd_2 = Command::move_call(package_id, module, function_2, type_args, call_args_2);
+    pt_builder.command(cmd_1);
+    pt_builder.command(cmd_2);
+    let pt = pt_builder.finish();
+
+    let tx_data = TransactionData::new_programmable(signer, vec![gas], pt, 10_000_000, 1000);
+    let signed_data = cluster.wallet.sign_transaction(&tx_data);
+    let _response = client
+        .quorum_driver_api()
+        .execute_transaction_block(
+            signed_data,
+            IotaTransactionBlockResponseOptions::new(),
+            Some(ExecuteTransactionRequestType::WaitForLocalExecution),
+        )
+        .await
+        .unwrap();
+    // match with None function, the DB should have 2 records, but both points to
+    // the same tx
+    let filter = TransactionFilter::MoveFunction {
+        package: package_id,
+        module: Some("pay".to_string()),
+        function: None,
+    };
+    let move_call_query = IotaTransactionBlockResponseQuery::new_with_filter(filter);
+    let tx = client
+        .read_api()
+        .query_transaction_blocks(move_call_query, None, Some(20), true)
+        .await
+        .unwrap();
+    // verify that only 1 tx is returned and no
+    // IotaRpcInputError::ContainsDuplicates error
+    assert_eq!(1, tx.data.len());
+    Ok(())
+}
+
+#[sim_test]
+async fn test_add_dynamic_field() -> Result<(), anyhow::Error> {
+    let mut cluster = TestClusterBuilder::new().build().await;
+    let context = &cluster.wallet;
+    let client = context.get_client().await.unwrap();
+    let rpc_client = cluster.rpc_client();
+    let signer = cluster.get_address_0();
+
+    // Get the first object owned by the signer (to act as parent UID object)
+    let address = cluster.get_address_0();
+    let objects = context
+        .get_all_gas_objects_owned_by_address(address)
+        .await
+        .unwrap();
+
+    let parent_object = objects.first().unwrap(); // Get parent object (UID)
+    let child_object = &objects[1]; // Get child object
+    let gas = objects.last().unwrap(); // Get gas object
+
+    // Define the module and function details
+    let package_id = ObjectID::new(IOTA_FRAMEWORK_ADDRESS.into_bytes());
+    let module = Identifier::from_str("dynamic_field")?;
+    let function = Identifier::from_str("add")?;
+
+    // Build the transaction
+    let mut pt_builder = ProgrammableTransactionBuilder::new();
+
+    // Parent object must be passed as mutable reference (UID)
+    let parent_argument = pt_builder
+        .obj(ObjectArg::ImmOrOwnedObject(*parent_object))
+        .expect("valid obj");
+
+    // Field name and field value as pure u64 types
+    let field_name_argument = pt_builder.pure(0).expect("valid pure");
+    let field_value_argument = pt_builder.pure(0).expect("valid pure");
+
+    // Add the command to call the `add` function of the dynamic_field module
+    let cmd = Command::move_call(
+        package_id,
+        module,
+        function,
+        vec![TypeTag::U64, TypeTag::U64],
+        vec![parent_argument, field_name_argument, field_value_argument],
+    );
+    pt_builder.command(cmd);
+
+    let pt = pt_builder.finish();
+
+    // Finish building the transaction
+    let tx_builder = TestTransactionBuilder::new(signer, *gas, 1000);
+    let txn = context.sign_transaction(&tx_builder.programmable(pt).build());
+
+    let resp = context.execute_transaction_must_succeed(txn).await;
+
+    // Sign and execute the transaction
+    tokio::time::sleep(tokio::time::Duration::from_secs(10)).await;
+
+    // Verify that the dynamic field was successfully added
+    let dynamic_fields = rpc_client
+        .get_dynamic_fields(parent_object.0, None, None)
+        .await
+        .expect("Failed to get dynamic fields");
+
+    assert!(
+        !dynamic_fields.data.is_empty(),
+        "Dynamic field was not added"
+    );
 
     Ok(())
 }
