+++ crates/iota-json-rpc-tests/tests/transaction_tests.rs
@@ -157,293 +157,3 @@ async fn test_get_raw_transaction() -> Result<(), anyhow::Error> {
 
     Ok(())
 }
-
-#[sim_test]
-async fn test_get_fullnode_transaction() -> Result<(), anyhow::Error> {
-    let cluster = TestClusterBuilder::new().build().await;
-
-    let context = &cluster.wallet;
-
-    let mut tx_responses: Vec<IotaTransactionBlockResponse> = Vec::new();
-
-    let client = context.get_client().await.unwrap();
-
-    for address in cluster.get_addresses() {
-        let objects = client
-            .read_api()
-            .get_owned_objects(
-                address,
-                Some(IotaObjectResponseQuery::new_with_options(
-                    IotaObjectDataOptions::new()
-                        .with_type()
-                        .with_owner()
-                        .with_previous_transaction(),
-                )),
-                None,
-                None,
-            )
-            .await?
-            .data;
-        let gas_id = objects.last().unwrap().object().unwrap().object_id;
-
-        // Make some transactions
-        for obj in &objects[..objects.len() - 1] {
-            let oref = obj.object().unwrap();
-            let data = client
-                .transaction_builder()
-                .transfer_object(address, oref.object_id, Some(gas_id), 1_000_000, address)
-                .await?;
-            let tx = cluster.wallet.sign_transaction(&data);
-
-            let response = client
-                .quorum_driver_api()
-                .execute_transaction_block(
-                    tx,
-                    IotaTransactionBlockResponseOptions::new(),
-                    Some(ExecuteTransactionRequestType::WaitForLocalExecution),
-                )
-                .await
-                .unwrap();
-
-            tx_responses.push(response);
-        }
-    }
-
-    // test get_recent_transactions with smaller range
-    let query = IotaTransactionBlockResponseQuery {
-        options: Some(IotaTransactionBlockResponseOptions {
-            show_input: true,
-            show_effects: true,
-            show_events: true,
-            ..Default::default()
-        }),
-        ..Default::default()
-    };
-
-    let tx = client
-        .read_api()
-        .query_transaction_blocks(query, None, Some(3), true)
-        .await
-        .unwrap();
-    assert_eq!(3, tx.data.len());
-    assert!(tx.data[0].transaction.is_some());
-    assert!(tx.data[0].effects.is_some());
-    assert!(tx.data[0].events.is_some());
-    assert!(tx.has_next_page);
-
-    // test get all transactions paged
-    let first_page = client
-        .read_api()
-        .query_transaction_blocks(
-            IotaTransactionBlockResponseQuery::default(),
-            None,
-            Some(5),
-            false,
-        )
-        .await
-        .unwrap();
-    assert_eq!(5, first_page.data.len());
-    assert!(first_page.has_next_page);
-
-    let second_page = client
-        .read_api()
-        .query_transaction_blocks(
-            IotaTransactionBlockResponseQuery::default(),
-            first_page.next_cursor,
-            None,
-            false,
-        )
-        .await
-        .unwrap();
-    assert!(second_page.data.len() > 5);
-
-    let mut all_txs = first_page.data.clone();
-    all_txs.extend(second_page.data);
-
-    // test get 10 transactions paged
-    let latest = client
-        .read_api()
-        .query_transaction_blocks(
-            IotaTransactionBlockResponseQuery::default(),
-            None,
-            Some(10),
-            false,
-        )
-        .await
-        .unwrap();
-    assert_eq!(10, latest.data.len());
-    assert_eq!(Some(all_txs[9].digest), latest.next_cursor);
-    assert_eq!(all_txs[0..10], latest.data);
-    assert!(latest.has_next_page);
-
-    // test get from address txs in ascending order
-    let address_txs_asc = client
-        .read_api()
-        .query_transaction_blocks(
-            IotaTransactionBlockResponseQuery::new_with_filter(TransactionFilter::FromAddress(
-                cluster.get_address_0(),
-            )),
-            None,
-            None,
-            false,
-        )
-        .await
-        .unwrap();
-    assert_eq!(4, address_txs_asc.data.len());
-
-    // test get from address txs in descending order
-    let address_txs_desc = client
-        .read_api()
-        .query_transaction_blocks(
-            IotaTransactionBlockResponseQuery::new_with_filter(TransactionFilter::FromAddress(
-                cluster.get_address_0(),
-            )),
-            None,
-            None,
-            true,
-        )
-        .await
-        .unwrap();
-    assert_eq!(4, address_txs_desc.data.len());
-
-    // test get from address txs in both ordering are the same.
-    let mut data_asc = address_txs_asc.data;
-    data_asc.reverse();
-    assert_eq!(data_asc, address_txs_desc.data);
-
-    // test get_recent_transactions
-    let tx = client
-        .read_api()
-        .query_transaction_blocks(
-            IotaTransactionBlockResponseQuery::default(),
-            None,
-            Some(20),
-            true,
-        )
-        .await
-        .unwrap();
-    assert_eq!(20, tx.data.len());
-
-    // test get_transaction
-    for tx_resp in tx.data {
-        let response: IotaTransactionBlockResponse = client
-            .read_api()
-            .get_transaction_with_options(
-                tx_resp.digest,
-                IotaTransactionBlockResponseOptions::new(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(tx_resp.digest, response.digest);
-    }
-
-    Ok(())
-}
-
-#[sim_test]
-async fn test_query_transaction_blocks() -> Result<(), anyhow::Error> {
-    let mut cluster = TestClusterBuilder::new().build().await;
-    let context = &cluster.wallet;
-    let client = context.get_client().await.unwrap();
-
-    let address = cluster.get_address_0();
-    let objects = client
-        .read_api()
-        .get_owned_objects(
-            address,
-            Some(IotaObjectResponseQuery::new_with_options(
-                IotaObjectDataOptions::new()
-                    .with_type()
-                    .with_owner()
-                    .with_previous_transaction(),
-            )),
-            None,
-            None,
-        )
-        .await?
-        .data;
-
-    // make 2 move calls of same package & module, but different functions
-    let package_id = ObjectID::new(IOTA_FRAMEWORK_ADDRESS.into_bytes());
-    let coin = objects.first().unwrap();
-    let coin_2 = &objects[1];
-    let signer = cluster.wallet.active_address().unwrap();
-
-    let tx_builder = client.transaction_builder().clone();
-    let mut pt_builder = ProgrammableTransactionBuilder::new();
-    let gas = objects.last().unwrap().object().unwrap().object_ref();
-
-    let module = Identifier::from_str("pay")?;
-    let function_1 = Identifier::from_str("split")?;
-    let function_2 = Identifier::from_str("divide_and_keep")?;
-
-    let iota_type_args = type_args![GAS::type_tag()]?;
-    let type_args = iota_type_args
-        .into_iter()
-        .map(|ty| ty.try_into())
-        .collect::<Result<Vec<_>, _>>()?;
-
-    let iota_call_args_1 = call_args!(coin.data.clone().unwrap().object_id, 10)?;
-    let call_args_1 = tx_builder
-        .resolve_and_checks_json_args(
-            &mut pt_builder,
-            package_id,
-            &module,
-            &function_1,
-            &type_args,
-            iota_call_args_1,
-        )
-        .await?;
-    let cmd_1 = Command::move_call(
-        package_id,
-        module.clone(),
-        function_1,
-        type_args.clone(),
-        call_args_1.clone(),
-    );
-
-    let iota_call_args_2 = call_args!(coin_2.data.clone().unwrap().object_id, 10)?;
-    let call_args_2 = tx_builder
-        .resolve_and_checks_json_args(
-            &mut pt_builder,
-            package_id,
-            &module,
-            &function_2,
-            &type_args,
-            iota_call_args_2,
-        )
-        .await?;
-    let cmd_2 = Command::move_call(package_id, module, function_2, type_args, call_args_2);
-    pt_builder.command(cmd_1);
-    pt_builder.command(cmd_2);
-    let pt = pt_builder.finish();
-
-    let tx_data = TransactionData::new_programmable(signer, vec![gas], pt, 10_000_000, 1000);
-    let signed_data = cluster.wallet.sign_transaction(&tx_data);
-    let _response = client
-        .quorum_driver_api()
-        .execute_transaction_block(
-            signed_data,
-            IotaTransactionBlockResponseOptions::new(),
-            Some(ExecuteTransactionRequestType::WaitForLocalExecution),
-        )
-        .await
-        .unwrap();
-    // match with None function, the DB should have 2 records, but both points to
-    // the same tx
-    let filter = TransactionFilter::MoveFunction {
-        package: package_id,
-        module: Some("pay".to_string()),
-        function: None,
-    };
-    let move_call_query = IotaTransactionBlockResponseQuery::new_with_filter(filter);
-    let tx = client
-        .read_api()
-        .query_transaction_blocks(move_call_query, None, Some(20), true)
-        .await
-        .unwrap();
-    // verify that only 1 tx is returned and no
-    // IotaRpcInputError::ContainsDuplicates error
-    assert_eq!(1, tx.data.len());
-    Ok(())
-}
